Detangle the OSM extraction process into the street_network part and everything else.
