+++ convert_osm/src/extract.rs
@@ -1,35 +1,15 @@
-use std::collections::{HashMap, HashSet};
-
-use osm::{NodeID, OsmID, RelationID, WayID};
+use osm::{OsmID, RelationID, WayID};
 
 use abstio::{CityName, MapName};
 use abstutil::{Tags, Timer};
-use geom::{Distance, FindClosest, HashablePt2D, Polygon, Pt2D, Ring};
+use geom::{Distance, FindClosest, Polygon, Pt2D, Ring};
 use kml::{ExtraShape, ExtraShapes};
-use raw_map::{
-    osm, Amenity, AreaType, Direction, DrivingSide, NamePerLanguage, RawArea, RawBuilding, RawMap,
-    RawParkingLot, RestrictionType,
-};
+use raw_map::{Amenity, AreaType, RawArea, RawBuilding, RawMap, RawParkingLot};
+use street_network::{osm, NamePerLanguage};
 
+use crate::import_streets::OsmExtract;
 use crate::osm_reader::{get_multipolygon_members, glue_multipolygon, multipoly_geometry};
 use crate::Options;
-
-pub struct OsmExtract {
-    /// Unsplit roads. These aren't RawRoads yet, because they may not obey those invariants.
-    pub roads: Vec<(WayID, Vec<Pt2D>, Tags)>,
-    /// Traffic signals to the direction they apply
-    pub traffic_signals: HashMap<HashablePt2D, Direction>,
-    pub osm_node_ids: HashMap<HashablePt2D, NodeID>,
-    /// (ID, restriction type, from way ID, via node ID, to way ID)
-    pub simple_turn_restrictions: Vec<(RestrictionType, WayID, NodeID, WayID)>,
-    /// (relation ID, from way ID, via way ID, to way ID)
-    pub complicated_turn_restrictions: Vec<(RelationID, WayID, WayID, WayID)>,
-    /// Crosswalks located at these points, which should be on a RawRoad's center line
-    pub crosswalks: HashSet<HashablePt2D>,
-    /// Some kind of barrier nodes at these points. Only the ones on a RawRoad center line are
-    /// relevant.
-    pub barrier_nodes: HashSet<HashablePt2D>,
-}
 
 pub fn extract_osm(
     map: &mut RawMap,
@@ -52,6 +32,14 @@ pub fn extract_osm(
     if let Some(way) = doc.ways.get_mut(&WayID(332355467)) {
         way.tags.insert("junction", "intersection");
     }
+    if let Some(way) = doc.ways.get_mut(&WayID(332060260)) {
+        way.tags.insert("sidewalk", "right");
+    }
+    if let Some(way) = doc.ways.get_mut(&WayID(332060236)) {
+        way.tags.insert("sidewalk", "right");
+    }
+    // Hack for Geneva, which maps sidewalks as separate ways
+    let infer_both_sidewalks_for_oneways = map.name.city == CityName::new("ch", "geneva");
 
     if clip_path.is_none() {
         // Use the boundary from .osm.
@@ -59,37 +47,13 @@ pub fn extract_osm(
         map.streets.boundary_polygon = map.streets.gps_bounds.to_bounds().get_rectangle();
     }
 
-    let mut out = OsmExtract {
-        roads: Vec::new(),
-        traffic_signals: HashMap::new(),
-        osm_node_ids: HashMap::new(),
-        simple_turn_restrictions: Vec::new(),
-        complicated_turn_restrictions: Vec::new(),
-        crosswalks: HashSet::new(),
-        barrier_nodes: HashSet::new(),
-    };
+    let mut out = OsmExtract::new();
     let mut amenity_points = Vec::new();
 
     timer.start_iter("processing OSM nodes", doc.nodes.len());
     for (id, node) in &doc.nodes {
         timer.next();
-        out.osm_node_ids.insert(node.pt.to_hashable(), *id);
-
-        if node.tags.is(osm::HIGHWAY, "traffic_signals") {
-            let dir = if node.tags.is("traffic_signals:direction", "backward") {
-                Direction::Back
-            } else {
-                Direction::Fwd
-            };
-            out.traffic_signals.insert(node.pt.to_hashable(), dir);
-        }
-        if node.tags.is(osm::HIGHWAY, "crossing") {
-            out.crosswalks.insert(node.pt.to_hashable());
-        }
-        // TODO Any kind of barrier?
-        if node.tags.is("barrier", "bollard") {
-            out.barrier_nodes.insert(node.pt.to_hashable());
-        }
+        out.handle_node(*id, node);
         for amenity in get_bldg_amenities(&node.tags) {
             amenity_points.push((node.pt, amenity));
         }
@@ -108,15 +72,7 @@ pub fn extract_osm(
 
         way.tags.insert(osm::OSM_WAY_ID, id.0.to_string());
 
-        if is_road(&mut way.tags, opts, &map.name) {
-            // TODO Hardcoding these overrides. OSM is correct, these don't have
-            // sidewalks; there's a crosswalk mapped. But until we can snap sidewalks properly, do
-            // this to prevent the sidewalks from being disconnected.
-            if id == WayID(332060260) || id == WayID(332060236) {
-                way.tags.insert(osm::SIDEWALK, "right");
-            }
-
-            out.roads.push((id, way.pts.clone(), way.tags.clone()));
+        if out.handle_way(id, &way, opts, infer_both_sidewalks_for_oneways) {
             continue;
         } else if way.tags.is(osm::HIGHWAY, "service") {
             // If we got here, is_road didn't interpret it as a normal road
@@ -199,7 +155,9 @@ pub fn extract_osm(
         timer.next();
         let id = *id;
 
-        if let Some(area_type) = get_area_type(&rel.tags) {
+        if out.handle_relation(id, rel) {
+            continue;
+        } else if let Some(area_type) = get_area_type(&rel.tags) {
             if rel.tags.is("type", "multipolygon") {
                 for polygon in
                     glue_multipolygon(id, get_multipolygon_members(id, rel, &doc), Some(&boundary))
@@ -210,52 +168,6 @@ pub fn extract_osm(
                         polygon,
                         osm_tags: rel.tags.clone(),
                     });
-                }
-            }
-        } else if rel.tags.is("type", "restriction") {
-            let mut from_way_id: Option<WayID> = None;
-            let mut via_node_id: Option<NodeID> = None;
-            let mut via_way_id: Option<WayID> = None;
-            let mut to_way_id: Option<WayID> = None;
-            for (role, member) in &rel.members {
-                match member {
-                    OsmID::Way(w) => {
-                        if role == "from" {
-                            from_way_id = Some(*w);
-                        } else if role == "to" {
-                            to_way_id = Some(*w);
-                        } else if role == "via" {
-                            via_way_id = Some(*w);
-                        }
-                    }
-                    OsmID::Node(n) => {
-                        if role == "via" {
-                            via_node_id = Some(*n);
-                        }
-                    }
-                    OsmID::Relation(r) => {
-                        warn!("{} contains {} as {}", id, r, role);
-                    }
-                }
-            }
-            if let Some(restriction) = rel.tags.get("restriction") {
-                if let Some(rt) = RestrictionType::new(restriction) {
-                    if let (Some(from), Some(via), Some(to)) = (from_way_id, via_node_id, to_way_id)
-                    {
-                        out.simple_turn_restrictions.push((rt, from, via, to));
-                    } else if let (Some(from), Some(via), Some(to)) =
-                        (from_way_id, via_way_id, to_way_id)
-                    {
-                        if rt == RestrictionType::BanTurns {
-                            out.complicated_turn_restrictions.push((id, from, via, to));
-                        } else {
-                            warn!(
-                                "Weird complicated turn restriction \"{}\" from {} to {} via {}: \
-                                 {}",
-                                restriction, from, to, via, id
-                            );
-                        }
-                    }
                 }
             }
         } else if is_bldg(&rel.tags) {
@@ -351,167 +263,6 @@ pub fn extract_osm(
     timer.stop("find service roads crossing parking lots");
 
     (out, amenity_points)
-}
-
-fn is_road(tags: &mut Tags, opts: &Options, name: &MapName) -> bool {
-    if tags.is("area", "yes") {
-        return false;
-    }
-
-    // First deal with railways.
-    if tags.is("railway", "light_rail") {
-        return true;
-    }
-    if tags.is("railway", "rail") && opts.include_railroads {
-        return true;
-    }
-
-    let highway = if let Some(x) = tags.get(osm::HIGHWAY) {
-        if x == "construction" {
-            // What exactly is under construction?
-            if let Some(x) = tags.get("construction") {
-                x
-            } else {
-                return false;
-            }
-        } else {
-            x
-        }
-    } else {
-        return false;
-    };
-
-    if !vec![
-        "cycleway",
-        "footway",
-        "living_street",
-        "motorway",
-        "motorway_link",
-        "path",
-        "pedestrian",
-        "primary",
-        "primary_link",
-        "residential",
-        "secondary",
-        "secondary_link",
-        "service",
-        "steps",
-        "tertiary",
-        "tertiary_link",
-        "track",
-        "trunk",
-        "trunk_link",
-        "unclassified",
-    ]
-    .contains(&highway.as_ref())
-    {
-        return false;
-    }
-
-    if highway == "track" && tags.is("bicycle", "no") {
-        return false;
-    }
-
-    #[allow(clippy::collapsible_if)] // better readability
-    if (highway == "footway" || highway == "path" || highway == "steps")
-        && opts.map_config.inferred_sidewalks
-    {
-        if !tags.is_any("bicycle", vec!["designated", "yes", "dismount"]) {
-            return false;
-        }
-    }
-    if highway == "pedestrian"
-        && tags.is("bicycle", "dismount")
-        && opts.map_config.inferred_sidewalks
-    {
-        return false;
-    }
-
-    // Import most service roads. Always ignore driveways, golf cart paths, and always reserve
-    // parking_aisles for parking lots.
-    if highway == "service" && tags.is_any("service", vec!["driveway", "parking_aisle"]) {
-        // An exception -- keep driveways signed for bikes
-        if !(tags.is("service", "driveway") && tags.is("bicycle", "designated")) {
-            return false;
-        }
-    }
-    if highway == "service" && tags.is("golf", "cartpath") {
-        return false;
-    }
-    if highway == "service" && tags.is("access", "customers") {
-        return false;
-    }
-
-    // Not sure what this means, found in Seoul.
-    if tags.is("lanes", "0") {
-        return false;
-    }
-
-    if opts.skip_local_roads && osm::RoadRank::from_highway(highway) == osm::RoadRank::Local {
-        return false;
-    }
-
-    // It's a road! Now fill in some possibly missing data.
-
-    // If there's no parking data in OSM already, then assume no parking and mark that it's
-    // inferred.
-    if !tags.contains_key(osm::PARKING_LEFT)
-        && !tags.contains_key(osm::PARKING_RIGHT)
-        && !tags.contains_key(osm::PARKING_BOTH)
-        && !tags.is_any(osm::HIGHWAY, vec!["motorway", "motorway_link", "service"])
-        && !tags.is("junction", "roundabout")
-    {
-        tags.insert(osm::PARKING_BOTH, "no_parking");
-        tags.insert(osm::INFERRED_PARKING, "true");
-    }
-
-    // If there's no sidewalk data in OSM already, then make an assumption and mark that
-    // it's inferred.
-    if !tags.contains_key(osm::SIDEWALK) && opts.map_config.inferred_sidewalks {
-        tags.insert(osm::INFERRED_SIDEWALKS, "true");
-
-        if tags.contains_key("sidewalk:left") || tags.contains_key("sidewalk:right") {
-            // Attempt to mangle
-            // https://wiki.openstreetmap.org/wiki/Key:sidewalk#Separately_mapped_sidewalks_on_only_one_side
-            // into left/right/both. We have to make assumptions for missing values.
-            let right = !tags.is("sidewalk:right", "no");
-            let left = !tags.is("sidewalk:left", "no");
-            let value = match (right, left) {
-                (true, true) => "both",
-                (true, false) => "right",
-                (false, true) => "left",
-                (false, false) => "none",
-            };
-            tags.insert(osm::SIDEWALK, value);
-        } else if tags.is_any(osm::HIGHWAY, vec!["motorway", "motorway_link"])
-            || tags.is_any("junction", vec!["intersection", "roundabout"])
-            || tags.is("foot", "no")
-            || tags.is(osm::HIGHWAY, "service")
-            // TODO For now, not attempting shared walking/biking paths.
-            || tags.is_any(osm::HIGHWAY, vec!["cycleway", "pedestrian", "track"])
-        {
-            tags.insert(osm::SIDEWALK, "none");
-        } else if tags.is("oneway", "yes") {
-            if opts.map_config.driving_side == DrivingSide::Right {
-                tags.insert(osm::SIDEWALK, "right");
-            } else {
-                tags.insert(osm::SIDEWALK, "left");
-            }
-            if tags.is_any(osm::HIGHWAY, vec!["residential", "living_street"])
-                && !tags.is("dual_carriageway", "yes")
-            {
-                tags.insert(osm::SIDEWALK, "both");
-            }
-            // Hack for Geneva, which maps sidewalks as separate ways
-            if name.city == CityName::new("ch", "geneva") {
-                tags.insert(osm::SIDEWALK, "both");
-            }
-        } else {
-            tags.insert(osm::SIDEWALK, "both");
-        }
-    }
-
-    true
 }
 
 fn is_bldg(tags: &Tags) -> bool {
